py/discord-bot/grajo-bot/requirements.txt
discord.py
pytz
python-dotenv
requests
validators
asyncio
uuid
datetime
discord-py-slash-command
pynacl
discord